Split lib/module_interface into server & client files. - This should allow us to have a better interface for clients that depends on client things (like ImageBuffers).
