+++ demo_modules/balls/balls_client.js
@@ -16,7 +16,7 @@ limitations under the License.
 import {GOOGLE_COLORS, BALL_RADIUS} from './constants.js';
 import {CanvasSurface} from '/client/surface/canvas_surface.ts';
 import {NumberLerpInterpolator, ValueNearestInterpolator} from '/client/network/state_manager.ts';
-import {Client} from '/lib/module_interface.ts';
+import {Client} from '/client/modules/module_interface.ts';
 
 export function load(state, wallGeometry) {
   class BallsClient extends Client {
