+++ demo_modules/big_image_carousel/big_image_carousel_client.js
@@ -13,7 +13,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 ==============================================================================*/
 
-import { Client } from '/lib/module_interface.ts';
+import { Client } from '/client/modules/module_interface.ts';
 import {P5Surface} from '/client/surface/p5_surface.ts';
 
 export function load(wallGeometry, asset) {
