+++ demo_modules/chaos/chaos_client.js
@@ -15,7 +15,7 @@ limitations under the License.
 
 import {GOOGLE_COLORS} from './colors.js';
 import {CanvasSurface} from '../../client/surface/canvas_surface.ts';
-import {Client} from "../../lib/module_interface.ts";
+import {Client} from "../../client/modules/module_interface.ts";
 
 export function load(network, wallGeometry) {
   // Maximum value in a particular pixel.
