+++ demo_modules/chaos/chaos_server.js
@@ -14,7 +14,7 @@ limitations under the License.
 ==============================================================================*/
 
 import {GOOGLE_COLORS} from './colors.js';
-import {Server} from '../../lib/module_interface.ts';
+import {Server} from '../../server/modules/module_interface.ts';
 
 export function load(network, wallGeometry) {
   // Set to true to enable console spam!
