+++ demo_modules/chicago-brick-live/chicago-brick-live-client.js
@@ -13,7 +13,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 ==============================================================================*/
 import {CanvasSurface} from '/client/surface/canvas_surface.ts';
-import {Client} from '/lib/module_interface.ts';
+import {Client} from '/client/modules/module_interface.ts';
 
 export function load(debug, network, wallGeometry) {
   const HIGHLIGHT_COLORS = ['#3cba54', '#f4c20d', '#db3236', '#4885ed'];
