+++ demo_modules/gears/gears_client.js
@@ -3,7 +3,7 @@
 import {GOOGLE_COLORS, DARK_COLORS} from './colors.js';
 import {CanvasSurface} from '/client/surface/canvas_surface.ts';
 import {Rectangle} from '/lib/math/rectangle.ts';
-import {Client} from '/lib/module_interface.ts';
+import {Client} from '/client/modules/module_interface.ts';
 
 export function load(debug, state, wallGeometry) {
   // Initially start with two layers.
