+++ demo_modules/operational_tests/memory_debug.js
@@ -13,7 +13,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 ==============================================================================*/
 
-import { Client } from "/lib/module_interface.ts";
+import { Client } from "/client/modules/module_interface.ts";
 
 export function load(network) {
   // This is a no-op module that shows what is leaking in the framework when we
