+++ demo_modules/operational_tests/threejs_test.js
@@ -15,7 +15,7 @@ limitations under the License.
 
 import * as Three from "https://esm.sh/three@0.145.0";
 import {ThreeJsSurface} from '../../client/surface/threejs_surface.ts';
-import { Client } from '../../lib/module_interface.ts'; 
+import { Client } from '../../client/modules/module_interface.ts'; 
 
 export function load(wallGeometry) {
   class ThreeJsTestClient extends Client {
