+++ demo_modules/p5_game_of_life/p5_game_of_life_client.js
@@ -16,7 +16,7 @@ limitations under the License.
 import {Polygon} from '/lib/math/polygon2d.ts';
 import {NUM_ROWS, NUM_COLUMNS} from './constants.js';
 import {P5Surface} from '/client/surface/p5_surface.ts';
-import {Client} from '/lib/module_interface.ts';
+import {Client} from '/client/modules/module_interface.ts';
 
 export function load(wallGeometry, network, debug) {
