+++ demo_modules/p5_template/p5_template.js
@@ -13,7 +13,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 ==============================================================================*/
 
-import { Client } from "/lib/module_interface.ts";
+import { Client } from "/client/modules/module_interface.ts";
 
 export function load(wallGeometry, debug, P5Surface) {
   // p5 must be a P5.js instance.
