+++ demo_modules/slideshow/slideshow_server.js
@@ -30,7 +30,7 @@ import LoadLocalStrategy from './load_local.js';
 import LoadFromFlickrStrategy from './load_from_flickr.js';
 import FullscreenDisplayStrategy from './fullscreen_display_server.js';
 import {SizeLimitedCache} from '../../lib/size_limited_cache.ts';
-import {Server} from '../../lib/module_interface.ts';
+import {Server} from '../../server/modules/module_interface.ts';
 
 import fetch from 'node-fetch';
 import randomjs from 'random-js';
