+++ demo_modules/tiffany/tiffany_client.js
@@ -1,7 +1,7 @@
 import {P5Surface} from '/client/surface/p5_surface.ts';
 import {sub} from '/lib/math/vector2d.ts';
 import {Polygon} from '/lib/math/polygon2d.ts';
-import { Client } from '/lib/module_interface.ts';
+import { Client } from '/client/modules/module_interface.ts';
 
 // These control the thickness of the lines that are drawn, and are
 // determined by taking the log10 of the number of underlying image pixels
