+++ server/modules/module.ts
@@ -23,7 +23,7 @@ import { delay } from "../../lib/promise.ts";
 import { getGeo } from "../util/wall_geometry.ts";
 import { clients } from "../network/network.ts";
 import * as path from "https://deno.land/std@0.129.0/path/mod.ts";
-import { Server } from "../../lib/module_interface.ts";
+import { Server } from "../../server/modules/module_interface.ts";
 import { EmptyModuleDef, ModuleDef } from "./module_def.ts";
 import { easyLog } from "../../lib/log.ts";
 import inject from "../../lib/inject.ts";
